requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
deepface==0.0.89
filelock==3.13.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.4.0
gdown==5.1.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==21.2.0
h5py==3.11.0
idna==3.7
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
namex==0.0.8
networkx==3.3
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
retina-face==0.0.16
rich==13.7.1
rsa==4.9
scikit-image==0.23.1
scipy==1.13.0
six==1.16.0
soupsieve==2.5
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tifffile==2024.2.12
tk==0.1.0
tqdm==4.66.2
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
Werkzeug==3.0.2
wrapt==1.16.0
